verify: 1U shifts in gpioctl/makefs, csh GENHDRS, extra CodeQL TUs
Unwind-32 FAILED print_caps and makefs ilog2 are type-level 1<<31, same class as BN_set_params. bin/csh now emits the Makefile GENHDRS so contrib/tcsh analyzes 51/54 OK. CodeQL COPYIN-OK on kern_context/prot/resource.

+++ hbsd/src/usr.sbin/makefs/ffs/mkfs.c
@@ -875,7 +875,7 @@ ilog2(int val)
 	u_int n;
 
 	for (n = 0; n < sizeof(n) * CHAR_BIT; n++)
-		if (1 << n == val)
+		if ((int)(1U << n) == val)
 			return (n);
 	errx(1, "%s: %d is not a power of 2", __func__, val);
 }
